cleanup: splitting into seperate analyze file
